Rename home package to view
