+++ DualView/src/main/java/com/microsoft/device/display/samples/dualview/ui/view/RestaurantView.kt
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-package com.microsoft.device.display.samples.dualview.ui.home
+package com.microsoft.device.display.samples.dualview.ui.view
 
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
